+++ genapp-sqltutorial/genappsqltutorial-model/src/main/java/org/fundaciobit/genappsqltutorial/model/dao/IProductsManager.java
@@ -7,7 +7,7 @@ import org.fundaciobit.genapp.common.i18n.I18NException;
 public interface IProductsManager extends org.fundaciobit.genapp.common.query.ITableManager<Products, Long> {
 
 
-	public Products create( java.lang.String _productname_, java.lang.Long _supplierid_, java.lang.Long _categoryid_, java.lang.String _unit_, java.lang.Integer _price_) throws I18NException;
+	public Products create( java.lang.String _productname_, java.lang.Long _supplierid_, java.lang.Long _categoryid_, java.lang.String _unit_, java.lang.Double _price_) throws I18NException;
 
 	public Products findByPrimaryKey(long _productid_);
