+++ genapp-sqltutorial/genappsqltutorial-persistence/src/main/java/org/fundaciobit/genappsqltutorial/persistence/ProductsJPAManager.java
@@ -58,7 +58,7 @@ public class ProductsJPAManager
         return list.toArray(new Products[list.size()]);
     };
 
-    public synchronized Products create( java.lang.String _productname_, java.lang.Long _supplierid_, java.lang.Long _categoryid_, java.lang.String _unit_, java.lang.Integer _price_) throws I18NException {
+    public synchronized Products create( java.lang.String _productname_, java.lang.Long _supplierid_, java.lang.Long _categoryid_, java.lang.String _unit_, java.lang.Double _price_) throws I18NException {
         ProductsJPA __bean =  new ProductsJPA(_productname_,_supplierid_,_categoryid_,_unit_,_price_);
         return create(__bean);
     }
